# task3/repo-stat/api/docs/swagger.yaml
basePath: /
definitions:
  github_com_Friend-zva_golang-course-task3_repo-stat_api_internal_dto.GetInfoRepoOutput:
    properties:
      created_at:
        example: Fri, 20 Aug 2021 09:38:00 UTC
        type: string
      description:
        example: Homework for GoLang course 2026
        type: string
      forks:
        example: 45
        type: integer
      full_name:
        example: golang-course
        type: string
      stars:
        example: 52
        type: integer
    type: object
  github_com_Friend-zva_golang-course-task3_repo-stat_api_internal_dto.PingResponse:
    properties:
      services:
        items:
          $ref: "#/definitions/github_com_Friend-zva_golang-course-task3_repo-stat_api_internal_dto.ServiceStatus"
        type: array
      status:
        type: string
    type: object
  github_com_Friend-zva_golang-course-task3_repo-stat_api_internal_dto.ServiceStatus:
    properties:
      name:
        type: string
      status:
        type: string
    type: object
  internal_controller_http.ErrorResponse:
    properties:
      error:
        example: repo 'go-course' not found
        type: string
    type: object
host: localhost:28080
info:
  contact:
    name: API Support
    url: https://github.com/Friend-zva/golang-course-task3
  description: |-
    API Gateway for fetching and processing GitHub repository information.
    The system demonstrates a microservices workflow with API Gateway, Processor, and Collector services.
    Built with Clean Architecture principles and gRPC for internal communication.
  title: Repo Stat API
  version: 1.0.0
paths:
  /api/ping:
    get:
      description: |-
        Monitors the availability of the entire internal chain (Processor and Subscriber).
        Returns 200 OK if all services are reachable, or 503 Service Unavailable if any service is down.
      produces:
        - application/json
      responses:
        "200":
          description: Successful response (All services UP)
          schema:
            $ref: "#/definitions/github_com_Friend-zva_golang-course-task3_repo-stat_api_internal_dto.PingResponse"
        "503":
          description: Service Unavailable (System Degraded)
          schema:
            $ref: "#/definitions/github_com_Friend-zva_golang-course-task3_repo-stat_api_internal_dto.PingResponse"
      summary: Health Check & Service Discovery
      tags:
        - status
  /api/repositories/info:
    get:
      description:
        Get information about a GitHub repository such as name, description,
        stars, forks, and creation date
      parameters:
        - description: Full URL of the GitHub repository
          example: '"https://github.com/golang/go"'
          in: query
          name: url
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: Successful response with repository statistics
          schema:
            $ref: "#/definitions/github_com_Friend-zva_golang-course-task3_repo-stat_api_internal_dto.GetInfoRepoOutput"
        "400":
          description: Invalid URL format
          schema:
            $ref: "#/definitions/internal_controller_http.ErrorResponse"
        "404":
          description: Repository not found
          schema:
            $ref: "#/definitions/internal_controller_http.ErrorResponse"
        "500":
          description: Internal server error
          schema:
            $ref: "#/definitions/internal_controller_http.ErrorResponse"
        "502":
          description: External GitHub API error
          schema:
            $ref: "#/definitions/internal_controller_http.ErrorResponse"
      summary: Get repository info
      tags:
        - repository
swagger: "2.0"